Attempt to execute a pipeline (assuming it has local data, or is passed local data).

Attempt to execute a pipeline (assuming it has local data, or is passed local data).

execute_data_frame(optree, ..., env = parent.frame(), data = NULL,
  result_limit = NULL)

Arguments

optree
rquery relop pipeline.

...
force later arguments to bind by name.

env
environment to work in.

data
data.frame to evaluate.

result_limit
numeric if not null limit result to this many rows.

Value

executed pipleline or NULL if not executable.

Examples

winvector_temp_db_handle <- list(
  db = DBI::dbConnect(RSQLite::SQLite(), ":memory:")
)
RSQLite::initExtension(winvector_temp_db_handle$db)

d <- data.frame(AUC = 0.6, R2 = c(0.1, 0.2), D = NA, z = 2)
optree <- table_source("d", c("AUC", "R2", "D")) %.>%
        extend_nse(., c := sqrt(R2)) %.>%
  orderby(.,  rev_cols = "R2")

rquery_apply_to_data_frame(d, optree)

#>   AUC  R2  D         c
#> 1 0.6 0.2 NA 0.4472136
#> 2 0.6 0.1 NA 0.3162278

execute_data_frame(optree, data = d)

#>   AUC  R2  D         c
#> 1 0.6 0.2 NA 0.4472136
#> 2 0.6 0.1 NA 0.3162278

d %.>% optree

#>   AUC  R2  D         c
#> 1 0.6 0.2 NA 0.4472136
#> 2 0.6 0.1 NA 0.3162278

# run (and build result for) ad-hoc query
d %.>%
  extend_nse(., c := sqrt(R2)) %.>%
  orderby(., rev_cols = "R2") %.>%
  execute_data_frame(.)

#>   AUC  R2  D z         c
#> 1 0.6 0.2 NA 2 0.4472136
#> 2 0.6 0.1 NA 2 0.3162278

# print ad-hoc query (result only available for printing)
d %.>%
  extend_nse(., c := sqrt(R2)) %.>%
  orderby(., rev_cols = "R2")

#> Error in rsqlite_send_query(conn@ptr, statement): external pointer is not valid

DBI::dbDisconnect(winvector_temp_db_handle$db)
winvector_temp_db_handle <- NULL