Materialize an optree as a table.

Run the data query as a CREATE TABLE AS . Think of as a function that can be applied to relop trees, not as a component to place in pipelines.

materialize(optree, db, table_name = mk_tmp_name_source("rquery_mat"), ...,
  overwrite = FALSE, temporary = TRUE)

Arguments

optree
relop operation tree.

db
DBI connecton.

table_name
character, name of table to create.

...
force later arguments to bind by name.

overwrite
logical if TRUE drop an previous table.

temporary
logical if TRUE try to create a temporary table.

Value

table handle

Examples

my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
d <- dbi_copy_to(my_db, 'd',
                data.frame(AUC = 0.6, R2 = 0.2))
optree <- extend_se(d, c("v" := "AUC + R2", "x" := "pmax(AUC,v)"))
cat(format(optree))

#> table('d') %.>%
#>  extend(.,
#>   v := AUC + R2) %.>%
#>  extend(.,
#>   x := max(AUC v))

res <- materialize(optree, my_db, "example")
cat(format(res))

#> table('example')

sql <- to_sql(res, my_db)
cat(sql)

#> SELECT
#>  `example`.`AUC`,
#>  `example`.`R2`,
#>  `example`.`v`,
#>  `example`.`x`
#> FROM
#>  `example`

DBI::dbGetQuery(my_db, sql)

#>   AUC  R2   v   x
#> 1 0.6 0.2 0.8 0.8

DBI::dbDisconnect(my_db)