Wrap a non-SQL node.

Note: non-SQL nodes are allowed to delete/overwrite both both the incoming and outgoing tables, so do not point them to non-temporary structures. Also they tend to land all columns (losing narrowing optimization), so can be expensive and should be used sparingly.

non_sql_node(source, ..., f_db, f_df = NULL, incoming_table_name,
  outgoing_table_name, columns_produced, display_form,
  pass_using = FALSE, orig_columns = TRUE, temporary = TRUE,
  env = parent.frame())

Arguments

source
source to work from (data.frame or relop node)

...
force later arguments to bind by name

f_db
database implementation signature: f_db(db, incoming_table_name, outgoing_table_name) (db being a database handle)

f_df
data.frame implementation signature: f_df(data.frame) (NULL defaults to taking from database).

incoming_table_name
character, name of incoming table

outgoing_table_name
character, name of produced table

columns_produced
character, names of additional columns produced

display_form
character, how to print node

pass_using
logical, if TRUE (or if f_db is NULL) pass using column calculations through (else assume using all columns).

orig_columns
logical if TRUE select all original columns.

temporary
logical, if TRUE mark tables temporary.

env
environment to look to.

Value

sql node.

See also