rQuery.attr {rQuery}	R Documentation

Get the value of an attribute for the first element in the set of matched elements or set one or more attributes for every matched element.

Description

Get the value of an attribute for the first element in the set of matched elements or set one or more attributes for every matched element.

Arguments

rquery
rQueryResult container

...
a namded list of attribute values

Value

list of attribute values or a rQueryResult with the modified elements

Author(s)

jfolson

[Package rQuery version 0.1 Index]