rQuery.before {rQuery}	R Documentation

Insert content, specified by the parameter, before each element in the set of matched elements.

Description

Insert content, specified by the parameter, before each element in the set of matched elements.

Arguments

rquery
rQueryResult container

value
XML element, rQueryResult container, or character vector of html

Value

the modified rquery object

Author(s)

jfolson

[Package rQuery version 0.1 Index]