rQuery.class {rQuery}	R Documentation

Get the class value(s) for the first element in the set of matched elements or add a class for every matched element.

Description

Get the class value(s) for the first element in the set of matched elements or add a class for every matched element.

Arguments

rquery
rQueryResult container

class
a new class value

Value

list of class values or a rQueryResult with the modified elements

Author(s)

jfolson

[Package rQuery version 0.1 Index]