rQuery.find {rQuery}	R Documentation

Get the descendants of each element in the current set of matched elements, filtered by a selector, jQuery object, or element.

Description

Get the descendants of each element in the current set of matched elements, filtered by a selector, jQuery object, or element.

Arguments

rquery
rQueryResult container to search

query
CSS selector query

Value

XMLAbstractNode wrapped in a rQueryResult

Author(s)

jfolson

[Package rQuery version 0.1 Index]