rQuery.html {rQuery}	R Documentation

Get the HTML contents of the first element in the set of matched elements or set the HTML contents of every matched element.

Description

Get the HTML contents of the first element in the set of matched elements or set the HTML contents of every matched element.

Arguments

rquery
an rQueryResult object

value
optional XML object or html character vector.

Value

rQueryResult containing the modified rquery element or a character vector containing the html contents of the first element.

[Package rQuery version 0.1 Index]