rQuery.insertAfter {rQuery}	R Documentation

Insert every element in the set of matched elements after the target.

Description

Insert every element in the set of matched elements after the target.

Arguments

rquery
rQueryResult container

value
XML element, rQueryResult container, or character vector of html

Value

the modified rquery object

Author(s)

jfolson

[Package rQuery version 0.1 Index]