rQueryResult {rQuery}	R Documentation

rQuery S4 class for containing selections is a simple wrapper around list

Description

This class exists primarily to facilitate the chaining syntax of jQuery.

Author(s)

jfolson

[Package rQuery version 0.1 Index]