Execute pipeline treating pipe_left_arg as local data to be copied into database.

Execute pipeline treating pipe_left_arg as local data to be copied into database.

# S3 method for relop
wrapr_function(pipe_left_arg, pipe_right_arg, pipe_environment,
  pipe_name = NULL)

Arguments

pipe_left_arg
data.frame or DBI database connection

pipe_right_arg
rquery relop operation tree

pipe_environment
environment to execute in

pipe_name
name of pipling symbol

Value

data.frame

See also

Examples

if (requireNamespace("DBI", quietly = TRUE) && requireNamespace("RSQLite", quietly = TRUE)) {
  # set up example database and
  # db execution helper
  db <- DBI::dbConnect(RSQLite::SQLite(),
                       ":memory:")
  RSQLite::initExtension(db)
  winvector_temp_db_handle <- list(db = db)

  # operations pipeline/tree
  optree <- table_source("d", "x") %.>%
    extend_nse(., y = x*x)

  # wrapr dot pipe wrapr_function dispatch
  # causes this statment to apply optree
  # to d.
  data.frame(x = 1:3) %.>% optree %.>% print(.)

  # remote example
  rq_copy_to(db, "d",
              data.frame(x = 7:8),
              overwrite = TRUE,
              temporary = TRUE)

  # wrapr dot pipe wrapr_function dispatch
  # causes this statment to apply optree
  # to db.
  db %.>% optree %.>% print(.)

  # clean up
  rm(list = "winvector_temp_db_handle")
  DBI::dbDisconnect(db)
}

#>   x y
#> 1 1 1
#> 2 2 4
#> 3 3 9
#>   x  y
#> 1 7 49
#> 2 8 64