Make a list of assignments, applying many functions to many columns.

Make a list of assignments, applying each function to each column named. Intended to be used as an argument in extend_se() or project_se().

make_assignments(columns, funs, ..., sep = "_", prefix = TRUE)

Arguments

columns

character, vector of column names to take values from.

funs

character, names of functions to apply.

...

not used, forced later parameters to bind by name

sep

character, naming separator

prefix

logical, if TRUE place function names prior, else after in results.

Examples

assignments <- make_assignments(c('x', 'y'), c('mean', med = 'median'))
print(assignments)
#>      mean_x      mean_y       med_x       med_y 
#>   "mean(x)"   "mean(y)" "median(x)" "median(y)" 
ops <- mk_td('d', c('x', 'y')) %.>% project_se(., assignments)
cat(format(ops))
#> mk_td("d", c(
#>   "x",
#>   "y")) %.>%
#>  project(., mean_x := mean(x), mean_y := mean(y), med_x := median(x), med_y := median(y),
#>   groupby = c())